Move KeyExpansion and TransactionSigner to test-clients because they are only used there.
