+++ test-clients/src/main/java/com/hedera/services/legacy/regression/umbrella/FileServiceTest.java
@@ -41,7 +41,7 @@ import com.hederahashgraph.api.proto.java.TransactionID;
 import com.hederahashgraph.api.proto.java.TransactionReceipt;
 import com.hederahashgraph.api.proto.java.TransactionResponse;
 import com.hederahashgraph.builder.RequestBuilder;
-import com.hederahashgraph.builder.TransactionSigner;
+import com.hedera.services.legacy.client.util.TransactionSigner;
 import com.hederahashgraph.service.proto.java.FileServiceGrpc.FileServiceBlockingStub;
 import io.grpc.ManagedChannel;
 import io.grpc.Status;
