+++ test-clients/src/main/java/com/hedera/services/legacy/regression/umbrella/SmartContractServiceTest.java
@@ -46,7 +46,7 @@ import com.hederahashgraph.api.proto.java.TransactionReceipt;
 import com.hederahashgraph.api.proto.java.TransactionRecord;
 import com.hederahashgraph.api.proto.java.TransactionResponse;
 import com.hederahashgraph.builder.RequestBuilder;
-import com.hederahashgraph.builder.TransactionSigner;
+import com.hedera.services.legacy.client.util.TransactionSigner;
 import com.hederahashgraph.service.proto.java.SmartContractServiceGrpc;
 import com.swirlds.common.CommonUtils;
 import io.grpc.Status;
