+++ test-clients/src/main/java/com/hedera/services/legacy/regression/umbrella/TestHelperComplex.java
@@ -29,7 +29,7 @@ import com.hederahashgraph.api.proto.java.ResponseType;
 import com.hederahashgraph.api.proto.java.Timestamp;
 import com.hederahashgraph.api.proto.java.Transaction;
 import com.hederahashgraph.builder.RequestBuilder;
-import com.hederahashgraph.builder.TransactionSigner;
+import com.hedera.services.legacy.client.util.TransactionSigner;
 
 import java.util.ArrayList;
 import java.util.List;
